% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{accuracyMeasures}
\alias{accuracyMeasures}
\title{accuracyMeasures}
\format{\preformatted{ NULL
}}
\usage{
accuracyMeasures(confusionDF, threshold = NULL, name = "model")
}
\arguments{
\item{confusionDF}{a data.frame containing columns truePos, falsePos, trueNeg, falseNeg}

\item{threshold=NULL}{if confusionDF contains more than one row, choose a useful threshold}

\item{name="model"}{}
}
\value{
data.frame containing name, accuracy, precision, recall, and f1
}
\description{
accuracyMeasures
}
\keyword{datasets}

