% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{calculateAUC}
\alias{calculateAUC}
\title{calculateAUC}
\format{\preformatted{ NULL
}}
\usage{
calculateAUC(confusionDF)
}
\arguments{
\item{confusionDF}{a data.frame containing columns truePos, falsePos, trueNeg, falseNeg}
}
\value{
list containing AUC and a data.frame containing SENS and ASP
}
\description{
get area under the curve and a data.frame containing SENS and ASP (for ROC curve)
}
\keyword{datasets}

