% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{confusionMatrix}
\alias{confusionMatrix}
\title{confusionMatrix}
\format{\preformatted{ NULL
}}
\usage{
confusionMatrix(data, responseColumn, predictionColumn, logging = FALSE)
}
\arguments{
\item{data}{}

\item{responseColumn}{}

\item{predictionColumn}{}
}
\value{
confusion matrix
}
\description{
confusionMatrix
}
\keyword{datasets}

