% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{getColumns}
\alias{getColumns}
\title{getColumns}
\format{\preformatted{ NULL
}}
\usage{
getColumns(data, columns)
}
\arguments{
\item{data}{}

\item{columns}{}
}
\value{
data containing only specified columns
}
\description{
getColumns
}
\keyword{datasets}

