% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{getConfusionList}
\alias{getConfusionList}
\title{getConfusionList}
\format{\preformatted{ NULL
}}
\usage{
getConfusionList(data, responseColumn, predictionColumn, logging = FALSE)
}
\arguments{
\item{data}{}

\item{responseColumn}{column containing the response variable}

\item{predictionColumn}{column containing the predictions (boolean)}
}
\value{
list containing true positives, false positives, true negatives, and false negatives
}
\description{
get confusion list for boolean predictionColumn
}
\keyword{datasets}

