% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recommender.R
\name{getRecommendations}
\alias{getRecommendations}
\title{getRecommendations}
\usage{
getRecommendations(data, model, key, keyColumn, valueColumn, type,
  threshold = NULL, limit = NULL, logging = FALSE)
}
\arguments{
\item{data}{}

\item{model}{}

\item{key}{element for which recommendations should be created}

\item{keyColumn}{columnName for keys}

\item{valueColumn}{columnName for values}

\item{type}{'confidence' or 'lift'}

\item{threshold}{}

\item{limit}{return top limit recommendations}

\item{logging=FALSE}{}
}
\value{
DataFrame containing recommendations
}
\description{
getRecommendations
}

