% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{getThreshold}
\alias{getThreshold}
\title{getThreshold}
\format{\preformatted{ NULL
}}
\usage{
getThreshold(data, responseCol = "response", predCol = "rate", n = 100,
  logging = FALSE)
}
\arguments{
\item{data}{}

\item{responseColumn}{column containing the response variable}

\item{predictionColumn}{column containing the predictions (double)}

\item{n=100}{number of thresholds to investigate}

\item{logging=FALSE}{}
}
\value{
data.frame containing true positives, false positives, true negatives, and false negatives for each investigated threshold and the optimal threshold
}
\description{
get optimal threshold for given predictions
}
\keyword{datasets}

