% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initializeSpark.R
\docType{data}
\name{initializeSpark}
\alias{initializeSpark}
\title{initializeSpark}
\format{\preformatted{ NULL
}}
\usage{
initializeSpark(sparkDirectory, size)
}
\arguments{
\item{sparkDirectory}{}

\item{size}{spark executor memory}
}
\value{
list(sc, sqlContext) sparkContext and sqlContext
}
\description{
initializeSpark
}
\keyword{datasets}

