% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KMeans.R
\docType{data}
\name{kmeans}
\alias{kmeans}
\title{kmeans}
\format{\preformatted{ NULL
}}
\usage{
kmeans(data, numericVars, responseColumn, k, eps = 0.01, maxIt = 10,
  logging = FALSE)
}
\arguments{
\item{data}{}

\item{numericVars}{}

\item{responseColumn}{}

\item{k}{number of clusters}

\item{eps}{= 0.01 stopping criterium}

\item{maxIt}{= 10 maximum number of iterations to run}
}
\value{
list data, scales, means
}
\description{
kmeans
}
\keyword{datasets}

