% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\docType{data}
\name{NaiveBayes}
\alias{NaiveBayes}
\title{NaiveBayes}
\format{\preformatted{ NULL
}}
\usage{
NaiveBayes(data, vars, nrOfNumericCategories = 100)
}
\arguments{
\item{data}{}

\item{vars}{Variables to create model}

\item{nrOfNumericCategories=100}{}
}
\description{
NaiveBayes
}
\keyword{datasets}

