% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{plotROC}
\alias{plotROC}
\title{plotROC}
\format{\preformatted{ NULL
}}
\usage{
plotROC(ROCdata)
}
\arguments{
\item{ROCdata}{a data.frame containing SENS and ASP}
}
\description{
plot ROC-curve
}
\keyword{datasets}

