% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KMeans.R
\docType{data}
\name{predict_kmeans}
\alias{predict_kmeans}
\title{predict_kmeans}
\format{\preformatted{ NULL
}}
\usage{
predict_kmeans(data, scales, means, responseColumn = "")
}
\arguments{
\item{data}{}

\item{scales}{}

\item{means}{}

\item{responseColumn=""}{}
}
\value{
data including prediction column 'cluster'
}
\description{
predict_kmeans
}
\keyword{datasets}

