% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\docType{data}
\name{predict_NaiveBayes}
\alias{predict_NaiveBayes}
\title{predict_NaiveBayes}
\format{\preformatted{ NULL
}}
\usage{
predict_NaiveBayes(data, scales, model, logging = FALSE)
}
\arguments{
\item{data}{}

\item{scales}{Scales returned by NaiveBayes(data, vars, ...)}

\item{model}{Model returned by NaiveBayes(data, vars, ...)}

\item{logging=FALSE}{enable/disable logging}
}
\description{
predict_NaiveBayes
}
\keyword{datasets}

