% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recommender.R
\docType{data}
\name{recommenderModel}
\alias{recommenderModel}
\title{recommenderModel}
\format{\preformatted{ NULL
}}
\usage{
recommenderModel(data, key, value, supportThreshold, logging = FALSE)
}
\arguments{
\item{data}{}

\item{key}{columnName for keys}

\item{value}{columnName for values}

\item{supportThreshold}{}

\item{data}{}

\item{key}{columnName for keys}

\item{value}{columnName for values}

\item{supportThreshold}{}

\item{data}{}

\item{model}{}

\item{key}{element for which recommendations should be created}

\item{keyColumn}{columnName for keys}

\item{valueColumn}{columnName for values}

\item{type}{'confidence' or 'lift'}

\item{threshold}{}

\item{limit}{return top limit recommendations}

\item{logging=FALSE}{}
}
\value{
DataFrame containing model

DataFrame containing model

DataFrame containing recommendations
}
\description{
recommenderModel

recommenderModel

recommenderModel
}
\keyword{datasets}

