% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{removeColumns}
\alias{removeColumns}
\title{removeColumns}
\format{\preformatted{ NULL
}}
\usage{
removeColumns(data, columns)
}
\arguments{
\item{data}{}

\item{columns}{}
}
\value{
data without sprcified columns
}
\description{
removeColumns
}
\keyword{datasets}

