% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{setType}
\alias{setType}
\title{setType}
\format{\preformatted{ NULL
}}
\usage{
setType(data, columns, type)
}
\arguments{
\item{data}{}

\item{columns}{which columns to cast}

\item{type}{which type to cast to}
}
\value{
data with casted columns
}
\description{
setType
}
\keyword{datasets}

